### 2017-09-03 #389 @andrewheusser
@cMadan Thanks!  We are hoping it will be useful to others.  We are using it in the lab and it streamlines our analyses quite a bit.

Regarding reviewers, would it be possible for you to review it?  Given your background in psychology and memory it seems like the best fit out of the list of reviewers, but I wasn't sure if that would be possible given that you are the editor.  If not, here are 3 more that also seem like a good fit:

1. arokem
2. jkahn
3. richardlitt
