### 2017-09-05 #389 @cMadan
@andrewheusser, unfortunately, I can't serve as both editor and reviewer. I also just moved to another uni/country last week and wouldn't have time to adequately review the project myself.

@jkahn and @RichardLitt, would either of you be able to review this project?
