+++ quail/load.py
@@ -13,7 +13,8 @@ import dill
 import pickle
 import os
 
-def load(dbpath=None, recpath=None, remove_subs=None, wordpool=None, groupby=None, experiments=None):
+def load(dbpath=None, recpath=None, remove_subs=None, wordpool=None, groupby=None, experiments=None,
+    filters=None):
     '''
     Function that loads sql files generated by autoFR Experiment
     '''
@@ -51,6 +52,9 @@ def load(dbpath=None, recpath=None, remove_subs=None, wordpool=None, groupby=Non
         # parse each participant's datastring as json object
         # and take the 'data' sub-object
         data = [json.loads(part)['data'] for part in data if part is not None]
+
+        # remove duplicate subject data for debugXG82XV:debug7XPXQA
+        data[110] = data[110][348:]
 
         # insert uniqueid field into trialdata in case it wasn't added
         # in experiment:
@@ -258,8 +262,16 @@ def load(dbpath=None, recpath=None, remove_subs=None, wordpool=None, groupby=Non
     # read in stimulus library
     wordpool = pd.read_csv(wordpool)
 
+    # add custom filters
+    if filters:
+        filter_func = [adaptive_filter, experimenter_filter, experiments_filter] + filters
+    else:
+        filter_func = [adaptive_filter, experimenter_filter, experiments_filter]
+
     # load in dbs and convert to df, and filter
-    dfs = [db2df(db, filter_func=[adaptive_filter, experimenter_filter, experiments_filter]) for db in dbpath]
+    dfs = [db2df(db, filter_func=filter_func) for db in dbpath]
+    # dfs = [db2df(db, filter_func=[experiments_filter]) for db in dbpath]
+
 
     # concatenate the db files
     df = pd.concat(dfs)
