+++ setup.py
@@ -25,7 +25,7 @@ setup(
         "matplotlib>=1.5.1",
         "scipy>=0.17.1",
         "numpy>=1.10.4",
-        "pandas==0.18.1",
+        "pandas>=0.20.3",
         "sqlalchemy",
         "dill",
         "requests",
