+++ build/lib/quail/load.py
@@ -1,351 +1,0 @@
-from __future__ import division
-from sqlalchemy import create_engine, MetaData, Table
-import json
-import math
-import re
-import csv
-from itertools import izip_longest
-from collections import Counter
-import pandas as pd
-import numpy as np
-from .egg import Egg
-import dill
-import pickle
-import os
-
-def load(dbpath=None, recpath=None, remove_subs=None, wordpool=None, groupby=None, experiments=None,
-    filters=None):
-    '''
-    Function that loads sql files generated by autoFR Experiment
-    '''
-
-    assert (dbpath is not None), "You must specify a db file or files."
-    assert (recpath is not None), "You must specify a recall folder."
-    assert (wordpool is not None), "You must specify a wordpool file."
-    assert (experiments is not None), "You must specify a list of experiments"
-
-    ############################################################################
-    # subfunctions #############################################################
-
-    def db2df(db, filter_func=None):
-        '''
-        Loads db file and converts to dataframe
-        '''
-        db_url = "sqlite:///" + db
-        table_name = 'turkdemo'
-        data_column_name = 'datastring'
-
-        # boilerplace sqlalchemy setup
-        engine = create_engine(db_url)
-        metadata = MetaData()
-        metadata.bind = engine
-        table = Table(table_name, metadata, autoload=True)
-
-        # make a query and loop through
-        s = table.select()
-        rows = s.execute()
-
-        data = []
-        for row in rows:
-            data.append(row[data_column_name])
-
-        # parse each participant's datastring as json object
-        # and take the 'data' sub-object
-        data = [json.loads(part)['data'] for part in data if part is not None]
-
-        # remove duplicate subject data for debugXG82XV:debug7XPXQA
-        data[110] = data[110][348:]
-
-        # insert uniqueid field into trialdata in case it wasn't added
-        # in experiment:
-        for part in data:
-            for record in part:
-        #         print(record)
-                if type(record['trialdata']) is list:
-
-                    record['trialdata'] = {record['trialdata'][0]:record['trialdata'][1]}
-                record['trialdata']['uniqueid'] = record['uniqueid']
-
-        # flatten nested list so we just have a list of the trialdata recorded
-        # each time psiturk.recordTrialData(trialdata) was called.
-        def isNotNumber(s):
-            try:
-                float(s)
-                return False
-            except ValueError:
-                return True
-
-        data = [record['trialdata'] for part in data for record in part]
-
-        # filter out fields that we dont want using isNotNumber function
-        filtered_data = [{k:v for (k,v) in part.items() if isNotNumber(k)} for part in data]
-
-        # Put all subjects' trial data into a dataframe object from the
-        # 'pandas' python library: one option among many for analysis
-        data_frame = pd.DataFrame(filtered_data)
-
-        data_column_name = 'codeversion'
-
-        # boilerplace sqlalchemy setup
-        engine = create_engine(db_url)
-        metadata = MetaData()
-        metadata.bind = engine
-        table = Table(table_name, metadata, autoload=True)
-
-        # make a query and loop through
-        s = table.select()
-        rows = s.execute()
-
-        versions = []
-        version_dict = {}
-        for row in rows:
-            version_dict[row[0]]=row[data_column_name]
-
-        version_col = []
-        for idx,sub in enumerate(data_frame['uniqueid'].unique()):
-            for i in range(sum(data_frame['uniqueid']==sub)):
-                version_col.append(version_dict[sub])
-        data_frame['exp_version']=version_col
-
-        if filter_func:
-            for idx,filt in enumerate(filter_func):
[…]
