### 2017-09-06 4230309
updated example data to fix pandas bug
