### 2017-09-06 c79e185
fixed bug w 2 subjects in example data
