### 2017-09-06 07f2099
added long description to setup.py
