### 2017-09-06 24e8bb2
removing long description
