### 2017-09-06 16f9a06
changed license to MIT instead of loading the whole thing in setup.py
