+++ setup.py
@@ -2,22 +2,15 @@
 
 from setuptools import setup, find_packages
 
-
-with open('README.md') as f:
-    readme = f.read()
-
-with open('LICENSE') as f:
-    license = f.read()
-
 setup(
     name='quail',
     version='0.1.2',
     description='A python toolbox for analyzing and plotting free recall data',
-    long_description=readme,
+    long_description=' ',
     author='Contextual Dynamics Lab',
     author_email='contextualdynamics@gmail.com',
     url='https://github.com/ContextLab/quail',
-    license=license,
+    license='MIT',
     packages=find_packages(exclude=('tests', 'docs', 'paper')),
     include_package_data=True,
     install_requires=[
