### 2017-09-07 #389 @jdkoen
Unfortunately, I too am unable to review and test it out at this time. Will
be out of the country for a week and am booked solid with participants the
following week.

Best,

Josh

On Wed, Sep 6, 2017 at 6:09 AM, Christopher Madan <notifications@github.com>
wrote:

> Thanks, @RichardLitt <https://github.com/richardlitt>! I'll ping you if
> we still haven't found someone by then.
>
