+++ build/lib/quail/helpers.py
@@ -1,0 +1,250 @@
+#!/usr/bin/env python
+
+from __future__ import division
+import pandas as pd
+import numpy as np
+import pickle
+
+def list2pd(all_data, subjindex=None, listindex=None):
+    """
+    Makes multi-indexed dataframe of subject data
+
+    Parameters
+    ----------
+    all_data : list of lists of strings
+        strings are either all presented or all recalled items, in the order of presentation or recall
+        *should also work for presented / recalled ints and floats, if desired
+
+
+    Returns
+    ----------
+    subs_list_of_dfs : multi-indexed dataframe
+        dataframe of subject data (presented or recalled words/items), indexed by subject and list number
+        cell populated by the term presented or recalled in the position indicated by the column number
+
+    """
+    # set default index if it is not defined
+    # max_nlists = max(map(lambda x: len(x), all_data))
+    listindex = [[idx for idx in range(len(sub))] for sub in all_data] if not listindex else listindex
+    subjindex = [idx for idx,subj in enumerate(all_data)] if not subjindex else subjindex
+
+
+    def make_multi_index(listindex, sub_num):
+        return pd.MultiIndex.from_tuples([(sub_num,lst) for lst in listindex], names = ['Subject', 'List'])
+
+    listindex = list(listindex)
+    subjindex = list(subjindex)
+
+    subs_list_of_dfs = [pd.DataFrame(sub_data, index=make_multi_index(listindex[sub_num], subjindex[sub_num])) for sub_num,sub_data in enumerate(all_data)]
+
+    return pd.concat(subs_list_of_dfs)
+
+def format2tidy(df, subjname, listname, subjgroup, **attrs):
+
+    melted_df = pd.melt(df.T)
+    melted_df[subjname]=""
+    for idx,sub in enumerate(melted_df['Subject'].unique()):
+        melted_df.loc[melted_df['Subject']==sub,subjname]=subjgroup[idx]
+    if attrs['analysis_type'] in ['spc']:
+        base = list(df.columns)
+        melted_df['Position'] = base * int(melted_df.shape[0] / len(base))
+        melted_df.columns = ['Subject', listname, 'Proportion Recalled', subjname, 'Position']
+    elif attrs['analysis_type'] in ['pfr', 'pnr']:
+        base = list(df.columns)
+        melted_df['Position'] = base * int(melted_df.shape[0] / len(base))
+        melted_df.columns = ['Subject', listname, 'Probability of Recall: Position ' + str(attrs['n']), subjname, 'Position']
+    elif attrs['analysis_type'] is 'lagcrp':
+        base = range(int(-len(df.columns.values)/2),int(len(df.columns.values)/2)+1)
+        melted_df['Position'] = base * int(melted_df.shape[0] / len(base))
+        melted_df.columns = ['Subject', listname, 'Conditional Response Probability', subjname, 'Position']
+    elif attrs['analysis_type'] is 'fingerprint' or attrs['analysis_type'] is 'fingerprint_temporal':
+        base = list(df.columns.values)
+        melted_df['Feature'] = base * int(melted_df.shape[0] / len(base))
+        melted_df.columns = ['Subject', listname, 'Clustering Score', subjname, 'Feature']
+    elif attrs['analysis_type'] is 'accuracy':
+        melted_df.columns = ['Subject', listname, 'Accuracy', subjname]
+    elif attrs['analysis_type'] is 'temporal':
+        melted_df.columns = ['Subject', listname, 'Temporal Clustering Score', subjname]
+
+
+    return melted_df
+
+def recmat2egg(recmat, list_length=None):
+        """
+        Creates egg data object from zero-indexed recall matrix
+
+        Parameters
+        ----------
[…]
