### 2017-09-14 fb6a2e8
changed load_example_data function path
