### 2017-09-19 ca9728c
added brief readme for ZimaEtal17 data
