+++ quail/data/README.md
@@ -1,0 +1,3 @@
+# Example data
+
+This folder contains data for [Ziman et al. (2017)](https://psyarxiv.com/psh48).  A description of the data format, along with documentation for loading in the data and carrying out the analyses from the paper may be found [here](http://cdl-quail.readthedocs.io/en/latest/tutorial.html).
