### 2017-09-25 78ddae0
saving progress on new fingerprint code
