+++ quail/__init__.py
@@ -3,4 +3,5 @@ from .egg import Egg
 from .analysis import analyze
 from .plot import plot
 from .helpers import stack_eggs, crack_egg, recmat2egg, load_egg
-from decode_speech import decode_speech
+from .decode_speech import decode_speech
+from .fingerprint import Fingerprint, OptimalPresenter
