### 2017-09-25 b554563
saving progress on new fingerprint code
