### 2017-09-25 cb6e731
saving progress on new fingerprint implementation
