+++ quail/fingerprint.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
 import numpy as np
 import warnings
-from .helpers import default_dist_funcs
+from .helpers import default_dist_funcs, parse_egg
+from .analysis import analyze_chunk, fingerprint_helper
 
 class Fingerprint(object):
     """
-    Class containing the memory fingerprint and associated functions
+    Class for the memory fingerprint and associated functions
 
     A memory fingerprint can be defined as a subject's tendency to cluster their
     recall responses with respect to more than one stimulus feature dimensions.
@@ -15,8 +16,11 @@ class Fingerprint(object):
     Parameters
     ----------
 
+    init : quail.Egg
+        Data to initialize the fingerprint instance
+
     features : list
-        Features to consider for fingerprint analyses
+        Features to consider for fingerprint analyses, defaults to all.
 
     state : np.array
         The current fingerprint (an array of real numbers between 0 and 1,
@@ -42,184 +46,61 @@ class Fingerprint(object):
         saved here.
     """
 
-    def __init__(self, features=None, state=None, n=0, permute=False, nperms=1000,
-                 dist_funcs=None, meta=None):
+    def __init__(self, init=None, features='all', state=None, n=0, permute=False, nperms=1000):
 
-        if state is None:
-            if features is not None:
-                self.state = np.array([.5 for feature in features])
-            else:
-                self.state = state
-
-        if meta is None:
-            meta = {}
-
-        # initialize history param
         self.history = []
-        self.history.append(self.state)
 
+        if init is not None:
+            data = analyze_chunk(init,
+                              analysis=fingerprint_helper,
+                              analysis_type='fingerprint',
+                              pass_features=True,
+                              parallel=False)
+            self.state = np.mean(data, 0)
+            self.features = data.columns.values.tolist()
+            self.history.append(self.state)
+            n+=1
+        else:
+            self.state = None
 
-        self.dist_funcs = dist_funcs
         self.n = n
 
-    def update(self, pres_list, rec_list, feature_list):
+    def update(self, egg):
         """
-        Updates fingerprint with new data
+        In-place method that updates fingerprint with new data
 
         Parameters
         ----------
-        pres_list : list
-            list of presented words
-        rec_list : list
-            list of recalled words
-        feature_list : list
-            list of feature dicts for presented words
-        dist_funcs : dict
-            dict of distance functions for each feature
-
+        egg : quail.Egg
+            Data to update fingerprint
         Returns
         ----------
         None
         """
 
-        # get distances
-        distances = self.compute_distances(pres_list, feature_list, self.dist_funcs)
-
-        # compute weights for a single list
-        next_weights = self.compute_feature_weights(pres_list, rec_list, feature_list, distances)
-
         # increment n
         self.n+=1
 
-        # multiply states by n
-        c = self.state*self.n
+        next_weights = np.mean(analyze_chunk(egg,
+                          analysis=fingerprint_helper,
+                          analysis_type='fingerprint',
+                          pass_features=True,
+                          parallel=False).as_matrix(), 0)
 
-        # update state
-        self.state = (c+next_weights)/(self.n+1)
+        if self.state is None:
+
+            # multiply states by n
+            c = self.state*self.n
+
+            # update state
+            self.state = (c+next_weights)/(self.n+1)
+
+        else:
+
+            self.state = next_weights
 
         # update the history
         self.history.append(next_weights)
-
-    def compute_distances(self, pres_list, feature_list, dist_funcs):
-        """
-        Compute distances between list words along n feature dimensions
-
-        Parameters
-        ----------
-        pres_list : list
-            list of presented words
-        feature_list : list
-            list of feature dicts for presented words
-        dist_funcs : dict
[…]
