### 2017-09-27 7e6ec11
finished stick breaking alg, still have to test
