### 2017-09-27 e39c700
finished stick breaking alg, still have to test
