+++ quail/fingerprint.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python
 import numpy as np
+from scipy.spatial.distance import cdist
 import warnings
+from .egg import Egg
 from .helpers import default_dist_funcs, parse_egg
-from .analysis import analyze_chunk, fingerprint_helper
+from .analysis import analyze_chunk, fingerprint_helper, compute_distances, compute_feature_weights
 
 class Fingerprint(object):
     """
@@ -102,6 +104,9 @@ class Fingerprint(object):
         # update the history
         self.history.append(next_weights)
 
+    def get_features(self):
+        return self.features
+
 class OptimalPresenter(object):
 
     def __init__(self, strategy='random', features=None, params=None,
@@ -136,25 +141,249 @@ class OptimalPresenter(object):
         """
         self.set_strategy = strategy
 
+    def order_perm(self, egg, nperms=10000):
+        """
+        This function re-sorts a list by computing permutations of a given
+        list and choosing the one that maximizes/minimizes variance.
+        """
+
+        # parse egg
+        pres, rec, features, dist_funcs = parse_egg(egg)
+
+        # compute distances
+        distances = compute_distances(pres, features, dist_funcs)
+
+        # length of list
+        pres_len = len(pres)
+
+        # initialize weights
+        weights = np.zeros((nperms, len(features[0].keys())))
+
+        # orders are nperms shuffled indexes
+        orders  = np.zeros((nperms, pres_len))
+
+        # loop over number of permuations
+        for perm in range(nperms):
+
+            # shuffle inds
+            idx = np.random.permutation(pres_len)
+
+            # shuffled pres
+            pres_perm = list(pres[idx])
+
+            # shuffled features
+            features_perm = list(features[idx])
+
+            # create a copy of distances
+            distances_perm = distances.copy()
+
+            for key in distances_perm:
+                distances_perm[key]=distances_perm[key][idx]
+
+            # compute weights
+            weights[perm,:] = compute_feature_weights(pres_perm, pres_perm, features_perm, distances_perm)
+
+            # save out the order
+            orders[perm,:] = idx
+
+        # get the fingerprint state
+        fingerprint = self.get_params('fingerprint').state
+
+        # find the closest
+        closest = orders[np.argmin(cdist(np.array(fingerprint, ndmin=2), weights)),:]
+
+        # return a re-sorted egg
+        return Egg(pres=[list(pres[idx])], rec=[[None]], features=[list(features[idx])])
+
+
     def order(self, egg):
         """
         Reorders a list according to strategy
         """
 
-        def get_feature_stick(w, alpha):
+        def compute_feature_stick(features, weights, alpha):
             '''create a 'stick' of feature weights'''
 
-            feature_stick = [[weights[feature]]*round(weights[feature]**alpha)*100 for feature in w]
-            return [item for sublist in feature_stick for item in sublist]
+            feature_stick = []
+            for f, w in zip(features, weights):
+                feature_stick+=[f]*int(np.power(w,alpha)*100)
 
-        def compute_stimulus_stick(s, tau):
-            '''create a 'stick' of feature weights'''
+            return feature_stick
 
-            feature_stick = [[weights[feature]]*round(weights[feature]**alpha)*100 for feature in w]
-            return [item for sublist in feature_stick for item in sublist]
+        def reorder_list(egg, feature_stick, tau):
+
+            def compute_stimulus_stick(s, tau):
+                '''create a 'stick' of feature weights'''
+
+                feature_stick = [[weights[feature]]*round(weights[feature]**alpha)*100 for feature in w]
+                return [item for sublist in feature_stick for item in sublist]
+
+            # parse egg
+            pres, rec, features, dist_funcs = parse_egg(egg)
+
[…]
