### 2017-10-02 ed0be88
saving progress on fingerprint, best stick not yet working...
