### 2017-10-02 30ffe9d
saving progress on fingerprint, best stick not yet working...
