+++ quail/fingerprint.py
@@ -2,13 +2,15 @@
 import numpy as np
 from scipy.spatial.distance import cdist
 import warnings
+from joblib import Parallel, delayed
+import multiprocessing
 from .egg import Egg
 from .helpers import default_dist_funcs, parse_egg
-from .analysis import analyze_chunk, fingerprint_helper, compute_distances, compute_feature_weights
+from .analysis import analyze_chunk, fingerprint_helper, compute_feature_weights
 
 class Fingerprint(object):
     """
-    Class for the memory fingerprint and associated functions
+    Class for the memory fingerprint
 
     A memory fingerprint can be defined as a subject's tendency to cluster their
     recall responses with respect to more than one stimulus feature dimensions.
@@ -108,6 +110,46 @@ class Fingerprint(object):
         return self.features
 
 class OptimalPresenter(object):
+    """
+    A class that reorders stimuli to optimize memory performance
+
+    A memory fingerprint can be defined as a subject's tendency to cluster their
+    recall responses with respect to more than one stimulus feature dimensions.
+    What is a 'stimulus feature dimension' you ask? It is simply an attribute of
+    the stimulus, such as its color, category, spatial location etc.
+
+    Parameters
+    ----------
+
+    init : quail.Egg
+        Data to initialize the fingerprint instance
+
+    features : list
+        Features to consider for fingerprint analyses, defaults to all.
+
+    state : np.array
+        The current fingerprint (an array of real numbers between 0 and 1,
+        inclusive) initialized to all 0.5
+
+    n : int
+        a counter specifying how many lists went into estimating the current
+        fingerprint (initialize to 0)
+
+    permute : bool
+        A boolean flag specifying whether to use permutations to compute the
+        fingerprint (default: True)
+
+
+    dist_funcs : dict (optional)
+        A dictionary of custom distance functions for stimulus features.  Each
+        key should be the name of a feature
+        and each value should be an inline distance function
+        (e.g. `dist_funcs['feature_n'] = lambda a, b: abs(a-b)`)
+
+    meta : dict (optional)
+        Meta data about the study (i.e. version, description, date, etc.) can be
+        saved here.
+    """
 
     def __init__(self, strategy='random', features=None, params=None,
                  fingerprint=None):
@@ -122,6 +164,8 @@ class OptimalPresenter(object):
         # update with user defined params
         if params is not None:
             self.params.update(params)
+
+        self.strategy = strategy
 
     def set_params(self, name, value):
         """
@@ -141,572 +185,402 @@ class OptimalPresenter(object):
         """
         self.set_strategy = strategy
 
-    def order_perm(self, egg, nperms=10000):
+    def order(self, egg, method='permute', nperms=10000, strategy=None):
         """
-        This function re-sorts a list by computing permutations of a given
-        list and choosing the one that maximizes/minimizes variance.
-        """
+        Reorders a list of stimuli to match a fingerprint
 
-        # parse egg
-        pres, rec, features, dist_funcs = parse_egg(egg)
+        Parameters
+        ----------
+        egg : quail.Egg
+            Data to compute fingerprint
 
-        # compute distances
-        distances = compute_distances(pres, features, dist_funcs)
+        method : str
+            Method to re-sort list. Can be 'stick' or 'permute' (default: permute)
 
-        # length of list
-        pres_len = len(pres)
+        nperms : int
+            Number of permutations to use. Only used if method='permute'. (default:
+            10,000)
 
-        # initialize weights
-        weights = np.zeros((nperms, len(features[0].keys())))
+        strategy : str or None
+            The strategy to use to reorder the list.  This can be 'stabilize',
+            'destabilize', 'random' or None.  If None, the self.strategy field
+            will be used. (default: None)
 
-        # orders are nperms shuffled indexes
[…]
