### 2017-10-04 #424 @whedon
[REVIEW]: Quail: A Python toolbox for analyzing and plotting free recall data
**Submitting author:** @andrewheusser (<a href="http://orcid.org/0000-0001-6353-688X">Andrew Heusser</a>)
**Repository:** <a href="https://github.com/ContextLab/quail" target ="_blank">https://github.com/ContextLab/quail</a>
**Version:** v0.1.0
