### 2017-10-04 aff5d60
saving progress on fingerprint code
