+++ quail/fingerprint.py
@@ -212,7 +212,7 @@ class OptimalPresenter(object):
             Egg re-sorted to match fingerprint
         """
 
-        def order_perm_parallel(self, egg, distances, strategy, nperms=10000):
+        def order_perm(self, egg, dist_dict, strategy, nperms=10000):
             """
             This function re-sorts a list by computing permutations of a given
             list and choosing the one that maximizes/minimizes variance.
@@ -225,7 +225,7 @@ class OptimalPresenter(object):
             pres_len = len(pres)
 
             results = Parallel(n_jobs=multiprocessing.cpu_count())(
-            delayed(run_perm)(pres, features, distances) for i in range(nperms))
+            delayed(rand_perm)(pres, features, dist_dict, dist_funcs) for i in range(nperms))
 
             weights = np.array(map(lambda x: x[0], results))
             orders = np.array(map(lambda x: x[1], results))
@@ -235,26 +235,20 @@ class OptimalPresenter(object):
 
             # find the closest (or farthest)
             if strategy is 'stabilize':
-                closest = orders[np.argmin(cdist(np.array(fingerprint, ndmin=2), weights)),:].astype(int).tolist()
+                closest = orders[np.argmin(cdist(np.array(fingerprint, ndmin=2), weights, 'correlation')),:].astype(int).tolist()
             elif strategy is 'destabilize':
-                closest = orders[np.argmax(cdist(np.array(fingerprint, ndmin=2), weights)),:].astype(int).tolist()
+                closest = orders[np.argmax(cdist(np.array(fingerprint, ndmin=2), weights, 'correlation')),:].astype(int).tolist()
 
             # return a re-sorted egg
             return Egg(pres=[list(pres[closest])], rec=[list(pres[closest])], features=[list(features[closest])])
 
-        def order_best_stick(self, egg, distances, strategy, nperms):
+        def order_best_stick(self, egg, dist_dict, strategy, nperms):
 
             # parse egg
             pres, rec, features, dist_funcs = parse_egg(egg)
 
-            # compute distances
-            # distances = compute_distances(pres, features, dist_funcs)
-
-            # length of list
-            pres_len = len(pres)
-
             results = Parallel(n_jobs=multiprocessing.cpu_count())(
-            delayed(stick_perm)(self, egg, distances, strategy) for i in range(nperms))
+            delayed(stick_perm)(self, egg, dist_dict, strategy) for i in range(nperms))
 
             weights = np.array(map(lambda x: x[0], results))
             orders = np.array(map(lambda x: x[1], results))
@@ -264,29 +258,57 @@ class OptimalPresenter(object):
 
             # find the closest (or farthest)
             if strategy is 'stabilize':
-                closest = orders[np.argmin(cdist(np.array(fingerprint, ndmin=2), weights)),:].astype(int).tolist()
+                closest = orders[np.argmin(cdist(np.array(fingerprint, ndmin=2), weights, 'correlation')),:].astype(int).tolist()
             elif strategy is 'destabilize':
-                closest = orders[np.argmax(cdist(np.array(fingerprint, ndmin=2), weights)),:].astype(int).tolist()
+                closest = orders[np.argmax(cdist(np.array(fingerprint, ndmin=2), weights, 'correlation')),:].astype(int).tolist()
 
             # return a re-sorted egg
-            return Egg(pres=[list(pres[closest])], rec=[list(pres[closest])], features=[list(features[closest])])
+            return Egg(pres=[list(pres[closest])], rec=[list(pres[closest])], features=[list(features[closest])], dist_funcs=dist_funcs)
+
+        def order_best_choice(self, egg, dist_dict, nperms):
+
+            # get the fingerprint state
+            fingerprint = self.get_params('fingerprint').state
+
[…]
