### 2017-10-04 1f85b2b
saving progress on fingerprint code
