### 2017-10-04 #424 @RichardLitt
@arfon You can do a request to https://api.github.com/repos/openjournals/joss-reviews/watchers to see if I am a watcher or not, before posting that message. Might be a good enhancement to @whedon.
