### 2017-10-04 c5d7697
added distfun arg to order function to allow reordering using different distance functions
