### 2017-10-04 fa043ed
added distfun arg to order function to allow reordering using different distance functions
