### 2017-10-04 ContextLab/quail#72 @RichardLitt
JOSS: Who is Campbell E. Field?
In the [paper.md](paper/paper.md) file, there is this listing:

```
 - name: Campbell E. Field
   orcid: 0000-0001-7260-635X
   affiliation: 1
```

Campbell does not seem to be a contributor of code to this repo. I am curious why that name was included in the paper. Related to https://github.com/openjournals/joss-reviews/issues/424.
