### 2017-10-04 ContextLab/quail#73 @RichardLitt
JOSS: Update submission version
Update the submission version for JOSS from v0.1.0 to v0.1.3, to match current releases.

Related to https://github.com/openjournals/joss-reviews/issues/424.
