### 2017-10-04 ContextLab/quail#72 @jeremymanning
@RichardLitt Campbell is one of the researchers who helped to test the software and collect experimental data that we've been using to test the toolbox.

The author roles were:
- Andrew Heusser: lead developer (code + design)
- Paxton Fitzpatrick, Campbell Field, and Kirsten Ziman: each contributed code and/or ideas and/or data
- Jeremy Manning: code, ideas, and overall design guidance
