### 2017-10-04 ContextLab/quail#72 @RichardLitt
Great. Just checking, then. Thank you!
