### 2017-10-04 7c92388
added alternative install instructions to readme
