### 2017-10-04 ContextLab/quail#75 @RichardLitt
Improve try it section - unable to launch binder or use Jupyter repo
Clicking on the binder link in the README led to http://mybinder.org/repo/contextlab/quail-example-notebooks, which presented a never-ending loading image. I am not sure that this is working as expected.

Likewise, going to the Jupyter repository was not useful. I don't know how to run Jupyter, or I am missing context on how to use that repository.

Part of https://github.com/openjournals/joss-reviews/issues/424.
