### 2017-10-04 ContextLab/quail#76 @RichardLitt
Unable to run example from README
The example in the README errored.

```
import quail

# presented words
presented_words = [['cat', 'bat', 'hat', 'goat'],['zoo', 'animal', 'zebra', 'horse']]

# recalled words
recalled_words = [['bat', 'cat', 'goat', 'hat'],['animal', 'horse', 'zoo']]

# create egg
egg = quail.Egg(pres=presented_words, rec=recalled_words)

#load data
egg = quail.load_example_data()

#analysis
analyzed_data = quail.analyze(egg, analysis='accuracy', listgroup=['average']*16)
```

I also tried with the second `egg` commented out. Should an argument be passed to `load_example_data()`? Would it be possible to provide an egg in the example which could be used to generate graphs as in the examples? Am I missing something?

Discovered as part of https://github.com/openjournals/joss-reviews/issues/424.
