### 2017-10-04 ContextLab/quail#75 @jeremymanning
Hi @RichardLitt, the binder server is down indefinitely, so we will remove that link in our readme file.

Here's a Jupyter tutorial: [[link](https://jupyter-notebook-beginner-guide.readthedocs.io/en/latest/)].  I'll add some documentation to that repository.
