### 2017-10-04 ContextLab/quail#77 @RichardLitt
documentation: Not all methods listed in readthedocs API documentation
For instance, `load.load_example_data` is not listed. Is there a reason only some methods are listed? That's the only one I could see in __init__.py which is relevant; of course, there are private methods in the python files which aren't exported.

On a related note, in the README Contributing section, it is specified that all public methods should be documented in the README. However, that is not where the documentation lives - it lives in docs, on readthedocs. Perhaps change the Contributing section?

I may be missing something here! Apologies if so. Part of https://github.com/openjournals/joss-reviews/issues/424.
