### 2017-10-04 ContextLab/quail#78 @RichardLitt
documentation: Add note about support
According to the Contribute section in the README, I can log a bug, and I can contribute. Where should I go for support? Can we add a Support section?

Part of https://github.com/openjournals/joss-reviews/issues/424.
