### 2017-10-04 ContextLab/quail#79 @RichardLitt
documentation: Outline key intended users
I would like to know _who_ this repository is made for. Right now, there are no intended users specified in the README or in the paper.md that I can tell. Can you please add a small description of the intended user?

Part of https://github.com/openjournals/joss-reviews/issues/424.
