### 2017-10-04 ContextLab/quail#80 @RichardLitt
JOSS: Add DOIs to all references which could have them
For instance, `@Hoga75`'s "Interitem encoding and directed search in free recall" should have one https://www.ncbi.nlm.nih.gov/pubmed/21287060 (10.3758/BF03212898).

Please double check the others and add where relevant.

Part of https://github.com/openjournals/joss-reviews/issues/424.
