### 2017-10-04 ContextLab/quail#75 @RichardLitt
Sounds good to me. Feel free to close when the README is updated.
