+++ paper/paper.bib
@@ -60,6 +60,7 @@
 
 @article{Hoga75,
 	author = {Hogan, R. M.},
+  doi = {10.3758/BF03212898},
 	journal = {MC},
 	pages = {197-209},
 	title = {Interitem encoding and directed search in free recall},
