### 2017-10-04 ContextLab/quail#80 @andrewheusser
@RichardLitt,  I added a DOI for the `@Hoga75` reference and checked the rest but they do not have one.

see: 183e922f5bef3dbfd450e65342fd98c55cc72b98
