### 2017-10-04 ContextLab/quail#79 @andrewheusser
@RichardLitt I've added a sentence in the README describing the intended user:

_The intended user of this toolbox is a memory researcher who seeks an easy way to analyze and visualize data from free recall psychology experiments._

see: 506283eaf126932515b5da12eee7dd8661269d81
