### 2017-10-04 ContextLab/quail#78 @andrewheusser
@RichardLitt I added a "support" section to the README on this commit: b00d3f8d5dca2238108687f44f862a7274491338
