### 2017-10-04 ContextLab/quail#76 @paxtonfitzpatrick
@RichardLitt should be fixed now
