### 2017-10-05 ContextLab/quail#79 @RichardLitt
Thanks. Could you add it to paper/paper.md too?
