### 2017-10-05 ContextLab/quail#76 @RichardLitt
Thank you. I am now able to run it.

The graph that appears is different in style from what is in the README. Is this expected?

![figure_1](https://user-images.githubusercontent.com/910753/31205552-0f6388fa-a940-11e7-92f6-736e7831eaa9.png)
