### 2017-10-05 ContextLab/quail#76 @paxtonfitzpatrick
Yes, that's expected for now.  I'll update the graphs that appear in the README to match the output shortly
