### 2017-10-05 f0fc4b8
added intended user sentence to paper.md
