### 2017-10-05 b9c5685
docstring/readthedocs for load_example_data, added test, fixed readme to point to readthedocs instead of readme
