+++ docs/quail.load_example_data.rst
@@ -1,0 +1,6 @@
+quail.load_example_data
+=======================
+
+.. currentmodule:: quail
+
+.. autofunction:: load_example_data
