### 2017-10-05 ContextLab/quail#77 @andrewheusser
@RichardLitt `quail.load_example_data` now has a docstring and is documented on readthedocs.  Also updated the README contribute section to point to the readthedocs instead of the README.  Thanks for the catch! b9c56857f67509b00509036f732befe3e9abf73f
