### 2017-10-05 ContextLab/quail#77 @RichardLitt
> quail.load_example_data now has a docstring and is documented on readthedocs.

I am unable to find this. What am I missing? Thanks!
