### 2017-10-05 ContextLab/quail#77 @andrewheusser
Sorry about that!  It worked locally, but the build failed on readthedocs.  Working now!

http://cdl-quail.readthedocs.io/en/latest/quail.load_example_data.html#quail.load_example_data
