### 2017-10-05 ContextLab/quail#75 @RichardLitt
Once this and #76 is landed, I think that's it for the JOSS review.
