### 2017-10-05 e5c4415
updated README figs
