### 2017-10-05 ContextLab/quail#76 @andrewheusser
@RichardLitt the figures are now updated to reflect the output of the README code (e5c441547faed527b054522271e7029ef4470c94)

cc: @paxtonfitzpatrick
