### 2017-10-05 e7cc70a
commented out binder badge until its working again
