### 2017-10-05 ContextLab/quail#75 @andrewheusser
@RichardLitt i commented out the binder link (I plan to uncomment it when binder is back online) and updated the `quail-example-notebooks` repo with instructions on how to get the notebooks running: https://github.com/ContextLab/quail-example-notebooks
