### 2017-10-05 ContextLab/quail#73 @jeremymanning
@andrewheusser can you tag the new version (0.1.4) and update [this file](https://github.com/ContextLab/quail/blob/master/paper/codemeta.json) once everything else is good to go with the submission (I think we've now dealt with everything-- @RichardLitt can you confim?).  That way the version of the software in the paper and on pip will include these latest changes to the documentation.
