### 2017-10-05 ContextLab/quail#73 @jeremymanning
(Also please push 0.1.4 to pip once it's tagged.)
