### 2017-10-05 ContextLab/quail#76 @RichardLitt
The new figures don't look the same as the one I posted above - no background, for instance. Do you have special settings for your output graphs?
