### 2017-10-05 ContextLab/quail#76 @andrewheusser
Ah! No, I think it stems from the fact that `seaborn`, a package that we rely on for plotting, updated their default figure styles.  I have an older version and you must have the newest.  I'll update the figures to reflect the latest version of seaborn.
