### 2017-10-05 ContextLab/quail#76 @andrewheusser
ok! figures are updated :)
