### 2017-10-05 #424 @RichardLitt
@cMadan I would consider Quail to be worthy of publication as is within JOSS. All checks passed, all green from my end.
