### 2017-10-05 #424 @RichardLitt
Small note: The only thing to change is that the version needs to be updated to v0.1.4. I'm not sure how to do that for JOSS.
