### 2017-10-05 ContextLab/quail#73 @andrewheusser
Ok, the version is updated and on pip: https://pypi.python.org/pypi/quail/0.1.4
