### 2017-10-05 ContextLab/quail#73 @jeremymanning
🎉  awesome!  closing this issue...
