### 2017-10-05 #424 @jeremymanning
@RichardLitt we've updated the metadata in the quail repository (and on pip) to reflect the new version (0.1.4).  Let us know if you need us to change anything else on our end, or if this is a JOSS thing.
