### 2017-10-05 #424 @jeremymanning
Also, thanks very much for such a thorough review!  An enjoyable experience from our end!
