### 2017-10-06 #424 @andrewheusser
Great, thank you!  Here is the DOI I registered with Zenodo

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1003184.svg)](https://doi.org/10.5281/zenodo.1003184)
