### 2017-10-07 #424 @jeremymanning
@arfon Thanks-- we're very excited that our paper has been accepted!  Everything mostly looks great, but I noticed some formatting issues with the references and paper.  I looked through our paper.md and paper.bib files, and I think I see what the problems were:
- In the .bib file, we had separated some of the author names with commas instead of "and."  Also, some of the field names were capitalized instead of lowercase.
- In the .md file, some of the parentheses placements make the formatting strange.

I've corrected these issues in [our repository](https://github.com/ContextLab/quail/tree/master/paper)-- is there any way for us to recompile the PDF to correct the formatting?
