### 2017-10-09 #424 @arfon
@jeremymanning - sure thing. I just recompiled the paper. Could you take a look at the new version and see if it's looking better?
