### 2017-10-09 #424 @jeremymanning
Hi @arfon -- that fixed some of the issues, but there are still some remaining formatting issues:
- Some misplaced parentheses
- Inconsistently formatted years (e.g. some have parentheses, some don't)
- Inconsistently formatted author names in the reference list, including some missing information

I've attached a marked up PDF with the specific issues.  I'm also happy to attempt to compile the PDF myself and/or generate a properly formatted one if you'd rather.  I could mock something up in LaTeX, or if you point me to the compiler you use, I can attempt to debug this.

[10.21105.joss.00424.pdf](https://github.com/openjournals/joss-reviews/files/1369318/10.21105.joss.00424.pdf)
