### 2017-10-09 #424 @arfon
@jeremymanning - [this is the command](https://github.com/openjournals/whedon/blob/d8edf88535db314874199e3ef4cc494e30ccec52/lib/whedon/processor.rb#L155-L164) we use to compile the paper. Can you go from there?
