### 2017-10-09 #424 @jeremymanning
Hi @arfon thanks for the pointer!  I've now fixed the remaining formatting issues.  The changes are [on our GitHub repo](https://github.com/ContextLab/quail/tree/master/paper), and I've attached a PDF.

[10.21105.joss.00424.pdf](https://github.com/openjournals/joss-reviews/files/1369829/10.21105.joss.00424.pdf)
