### 2017-10-09 #424 @arfon
@jeremymanning - can you take a look at this PDF https://github.com/openjournals/joss-papers/raw/master/joss.00424/10.21105.joss.00424.pdf and see if it looks better?
